git-svn-id: http://springhead.info/spr2/Springhead2/trunk@5767 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHHapticEngineImpulse.cpp
@@ -84,7 +84,7 @@ void PHHapticEngineImpulse::SyncPhysic2Haptic(){
 	for(int i = 0; i < NHapticPointers(); i++){
 		PHHapticPointer* ppointer = GetHapticPointer(i);
 		const int ppointerID = ppointer->GetPointerID();
-		for(int j = 0; j < ppointer->neighborSolidIDs.size(); j++){
+		for(size_t j = 0; j < ppointer->neighborSolidIDs.size(); j++){
 			const int solidID = ppointer->neighborSolidIDs[j];
 			PHSolidPairForHaptic* hpair = hapticLoop->GetSolidPairForHaptic(solidID, ppointerID);
 			PHSolidPairForHaptic* ppair = GetSolidPairForHaptic(solidID, ppointerID);
